requirements.txt
diffusers
transformers
accelerate
torch
torchvision
pillow
safetensors

